requirements.txt
langchain-groq
langchain-core
langgraph
fastapi
uvicorn
python-dotenv
pydantic
streamlit
requests